cami: Fix response splitting and premature login.
Fixes several major problems that caused C-AMI to
behave inconsistently.

The first two have to do with reading events from
the AMI socket. One possibility was that we would
finish reading from the socket without having received
the entire response. This would cause a partial
response to be sent to the consumer, which is bad.
The other possibility was the opposite, that we could
read multiple responses from the socket at once and
would therefore send multiple unrelated events to
a consumer at once, also bad. These issues have been
addressed by explicitly parsing each event received
and delivering events at the appropriate times.

Another unrelated issue was that, occasionally, we
might try to send the Login Action before Asterisk had
identified itself on the socket. This would cause login
to fail. We now prevent this by explicitly waiting for
Asterisk to ID itself before we attempt to log in.

With these changes, in informal testing that has been
conducted, the inconsistencies in how C-AMI can behave
seem to be resolved.