cami: fix bogus buffer size calculation
Previously, a buffer size calculation was using the length of the
format string and its arguments to determine the size, which is
wrong and led to format string truncation when sending AMI
actions. This is now fixed to correctly calculate using the actual
sizes of the buffer contents, including the initial format string
(but not its arguments, since they are not written into that buffer).