cami: prevent segfault on blank lines in AMI events
Blank lines in AMI events would result in an AMI field getting added
that didn't actually contain a key (let alone a value). Thus, calling
ami_keyvalue later would segfault when invoking a str comparison
function on the NULL key.
To prevent this, we now skip NULL keys in ami_keyvalue. We also
shift NULL fields to the end of an AMI event. We have to handle
in ami_keyvalue either way because we've already allocated the
space for all the fields, and they need to be freed, so we're just
going to have to put up with some nonused fields in this case due
to the malformed AMI event.