cami: fix response success return value.
Fixes "set" functions so that they return 0
on success and -1 on failure, not 1 for success
and 0 for failure. Also refactors this logic into
a separate function that users can call manually.
(This would generally be necessary only when using
the ami_action function directly, typically not
when using the built-in action functions that are
provided out of the box.)