cami: Fix handling of incomplete events
We currently handled the possibility of not receiving an entire response
in one socket read, but not the possibility of reading a partial event in
any given socket read. For example, it was possible to read part of an
event one read and then the rest of it the next. Because the while loop
for processing reads only executes when an AMI_EOM is present, nothing
would happen the first time and the second time, handling would start
from the beginning of what we read that particular read, as opposed to
at the beginning of the event which occurred the previous read. Among
other things, this could mess up the ability to detect the end of an event,
and the end of the event would not get picked up until there was further
data to read from Asterisk, and even then handling was not correct.

This is fixed by always starting from laststart instead of readinbuf, so that
partial events are handled properly.

Fixes #4.