cami: fix partial events being truncated and overwritten
If the beginning of a response is received in a read but the end
of the response "begin" event is not received until a future read,
then we don't currently handle that as any kind of ongoing
event so the beginning just gets overwritten which then messes
up all the processing afterwards. This fixes that by addressing the
opportunity where this could happen, by explicitly checking if
we have actually started to receive something and not overwriting
what we already received if that was the case.