cami: Fix handling of slow action responses.
Previously, if we did not receive a response to an action within
the threshold (1 second), we would abort and return -1.

This approach is problematic for two reasons:
* Some actions can legitimately take more than 1 second to return.
* Eventually we would get a response, but by that time we would
have already returned -1 to the user and now somebody else is
getting this event. Because we had assumed that the action IDs
would always match, this would cause this action to fail and essentially
cascade to a failure of all subsequent responses since they are all
essentially aligned improperly at that point.

To fix this, three changes have been made:
* The response wait timeout can now be overridden, and is currently
overridden by the Reload action, since some modules (e.g. PJSIP stuff)
can take longer to return a response.
* Instead of returning -1 if reading from the read pipe returns a different
Action ID, we continue waiting. This ensures that cascade failures don't happen.
This is to account for an insufficient waiting time, if and when that does happen,
though ideally this logic will not be used unless the timeout is too short.
* Logging improvements to provide more details about failures and anomalies.