cami: Fix duplicate connects for broken applications.
If an application tries to call ami_connect multiple
times, this can result in multiple AMI event loops
starting, which is very bad.

Although we previously had logic to close an existing
connection if the application did not, this is probably
not worth it. Instead, explicitly reject this and require
the application to close the AMI connection before doing
a reconnect, to avoid accidental duplicate connects.