libcami: Overhaul library to support multiple sessions.
Previously, CAMI only supported a single AMI connection,
using global variables for the connection. The library
has been modified to support multiple connections. Due
to the need for many functions to accept an argument
for the session, this is not ABI-compatible with previous
revisions of this library, so programs using CAMI will
need to be updated to compile.